refactor: сhange project structure
