+++ eslint.config.js
@@ -7,6 +7,7 @@ import tseslint from 'typescript-eslint';
 import eslintPluginPrettier from 'eslint-plugin-prettier/recommended';
 import reactCompiler from 'eslint-plugin-react-compiler';
 import simpleImportSort from 'eslint-plugin-simple-import-sort';
+import nextPlugin from '@next/eslint-plugin-next';
 
 export default tseslint.config(
   { ignores: ['eslint.config.js', 'vite.config.js', 'dist', 'coverage'] },
@@ -27,6 +28,7 @@ export default tseslint.config(
       'react-refresh': reactRefresh,
       'react-compiler': reactCompiler,
       'simple-import-sort': simpleImportSort,
+      '@next/next': nextPlugin,
     },
     rules: {
       ...reactHooks.configs.recommended.rules,
