+++ src/app/about/page.tsx
@@ -1,5 +1,27 @@
-import About from '@/pages/about/about';
+import Link from 'next/link';
+
+import './about.scss';
+
+import Section from '@/shared/ui/section/section';
+
+enum SchoolLinks {
+  SCHOOL = 'https://rs.school/',
+  REACT_CORSE = 'https://rs.school/courses/reactjs',
+}
 
 export default function AboutLayout() {
-  return <About />;
+  return (
+    <Section className="about">
+      <h1>My name is Sasha Hurbanava</h1>
+      <p>This is my study project for learning routings and hooks.</p>
+      <div className="links">
+        <Link href={SchoolLinks.SCHOOL} title="school">
+          <img src="/images.png"></img>
+        </Link>
+        <Link href={SchoolLinks.REACT_CORSE} title="react course">
+          <img src="/react.svg"></img>
+        </Link>
+      </div>
+    </Section>
+  );
 }
