+++ src/shared/components/root/root.scss
@@ -1,4 +1,4 @@
-@use '../../styles/scss/variables' as *;
+@use '../../../styles/scss/variables' as *;
 
 .theme {
   cursor: pointer;
