+++ src/shared/components/search-page/detailed/detailed-page.scss
