+++ src/shared/components/search-page/detailed/models/interfaces/character.ts
