+++ src/shared/components/search-page/detailed/models/interfaces/index.ts
