+++ src/shared/components/search-page/detailed/models/interfaces/selected-item.ts
