+++ src/shared/components/search-page/detailed/models/interfaces/writer.ts
