+++ src/shared/components/search-page/detailed/models/utils/index.ts
