+++ src/shared/components/search-page/detailed/models/utils/render-nested-object.tsx
