+++ src/shared/components/search-page/result-card/model/interfaces/card-props.ts
