+++ src/shared/components/search-page/result-card/model/interfaces/index.ts
