+++ src/shared/components/search-page/result-card/model/types/card-description.ts
