+++ src/shared/components/search-page/result-card/model/types/index.ts
