+++ src/shared/components/search-page/result-card/ui/result-card.tsx
