+++ src/shared/components/search-page/search-form/ui/search-form.scss
