+++ src/shared/components/search-page/search-page.scss
