+++ src/shared/components/search-page/search-page.tsx
