+++ src/shared/components/search-page/search-results/search-results.scss
